+++ Sources/VComponents/Components/Modals/Alert/VAlertExtension.swift
@@ -121,9 +121,7 @@ extension View {
         title: @escaping (Item) -> String?,
         message: @escaping (Item) -> String?,
         @VAlertButtonBuilder actions buttons: @escaping (Item) -> [any VAlertButtonProtocol]
-    ) -> some View
-        where Item: Identifiable
-    {
+    ) -> some View {
         item.wrappedValue.map { PresentationHostDataSourceCache.shared.set(key: id, value: $0) }
         
         return self
@@ -177,9 +175,7 @@ extension View {
         @ViewBuilder content: @escaping (Item) -> Content,
         @VAlertButtonBuilder actions buttons: @escaping (Item) -> [any VAlertButtonProtocol]
     ) -> some View
-        where
-            Item: Identifiable,
-            Content: View
+        where Content: View
     {
         item.wrappedValue.map { PresentationHostDataSourceCache.shared.set(key: id, value: $0) }