+++ Sources/VComponents/Helpers/Architectural Pattern Helpers/Alert/VAlertParameters.swift
@@ -10,59 +10,62 @@ import SwiftUI
 // MARK: - V Alert Parameters
 /// Parameters for presenting a `VAlert`.
 ///
-///     @State private var isPresented: Bool = false
+///     @State private var parameters: VAlertParameters?
 ///
 ///     var body: some View {
 ///         VPlainButton(
-///             action: { isPresented = true },
+///             action: {
+///                 parameters = VAlertParameters(
+///                     title: "Lorem Ipsum",
+///                     message: "Lorem ipsum dolor sit amet",
+///                     actions: {
+///                         VAlertButton(role: .primary, action: { print("Confirmed") }, title: "Confirm")
+///                         VAlertButton(role: .cancel, action: { print("Cancelled") }, title: "Cancel")
+///                     }
+///                 )
+///             },
 ///             title: "Present"
 ///         )
-///         .vAlert(
-///             id: "some_alert",
-///             isPresented: $isPresented,
-///             title: "Lorem Ipsum",
-///             message: "Lorem ipsum dolor sit amet",
-///             actions: {
-///                 VAlertButton(role: .primary, action: { print("Confirmed") }, title: "Confirm")
-///                 VAlertButton(role: .cancel, action: { print("Cancelled") }, title: "Cancel")
-///             }
-///         )
+///         .vAlert(id: "some_alert", parameters: $parameters)
 ///     }
-///     
+///
 @available(macOS, unavailable)
 @available(tvOS, unavailable)
 @available(watchOS, unavailable)
-public struct VAlertParameters: Identifiable {
+public struct VAlertParameters {
     // MARK: Properties
-    /// ID.
-    public let id: UUID = .init()
-    
     /// Title.
     public var title: String
     
     /// Message.
     public var message: String?
-    
+
     /// Buttons.
     public var buttons: () -> [any VAlertButtonProtocol]
+
+    /// Attributes.
+    public var attributes: [String: Any] = [:]
     
     // MARK: Parameters
     /// Initializes `VAlertParameters`.
     public init(
         title: String,
         message: String?,
-        @VAlertButtonBuilder actions buttons: @escaping () -> [any VAlertButtonProtocol]
+        @VAlertButtonBuilder actions buttons: @escaping () -> [any VAlertButtonProtocol],
+        attributes: [String: Any] = [:]
     ) {
         self.title = title
         self.message = message
         self.buttons = buttons
+        self.attributes = attributes
     }
     
     /// Initializes `VAlertParameters` with "ok" action.
     public init(
         title: String,
         message: String?,
-        completion: (() -> Void)?
+        completion: (() -> Void)?,
+        attributes: [String: Any] = [:]
     ) {
         self.init(
             title: title,
@@ -73,14 +76,16 @@ public struct VAlertParameters: Identifiable {
                     action: completion,
                     title: VComponentsLocalizationManager.shared.localizationProvider.vAlertCancelButtonTitle
                 )
-            }
+            },
+            attributes: attributes
         )
     }
     
     /// Initializes `VAlertParameters` with error and "ok" action.
     public init(
-        error: some Error,
-        completion: (() -> Void)?
+        error: any Error,
+        completion: (() -> Void)?,
+        attributes: [String: Any] = [:]
     ) {
         self.init(
             title: VComponentsLocalizationManager.shared.localizationProvider.vAlertErrorTitle,
@@ -91,7 +96,8 @@ public struct VAlertParameters: Identifiable {
                     action: completion,
                     title: VComponentsLocalizationManager.shared.localizationProvider.vAlertOKButtonTitle
                 )
-            }
+            },
+            attributes: attributes
         )
     }
 }